+++ modules/core/import/modules/csv/src/Plugin/Derivative/CsvImportMigrationLog.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\farm_import_csv\Plugin\Derivative;
 
+use Drupal\farm_flag\FarmFlagHelper;
 use Drupal\quantity\QuantityHelper;
 
 /**
@@ -54,6 +55,14 @@ class CsvImportMigrationLog extends CsvImportMigrationBase {
       }
     }
 
+    // Add allowed flags values.
+    foreach ($columns as &$column) {
+      if ($column['name'] == 'flags') {
+        $allowed_flags = FarmFlagHelper::flagOptions('log', [$bundle]);
+        $allowed_values_string = $this->t('Allowed values: @values.', ['@values' => implode(', ', array_keys($allowed_flags))]);
+        $column['description'] .= ' ' . $allowed_values_string;
+      }
+    }
   }
 
 }